db/src/main/java/be/ugent/degage/db/dao/AddressDAO.java
/* AddressDAO.java
 * ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 * Copyright Ⓒ 2014-2015 Universiteit Gent
 * 
 * This file is part of the Degage Web Application
 * 
 * Corresponding author (see also AUTHORS.txt)
 * 
 * Kris Coolsaet
 * Department of Applied Mathematics, Computer Science and Statistics
 * Ghent University 
 * Krijgslaan 281-S9
 * B-9000 GENT Belgium
 * 
 * The Degage Web Application is free software: you can redistribute it and/or modify
 * it under the terms of the GNU Affero General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 * 
 * The Degage Web Application is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU Affero General Public License for more details.
 * 
 * You should have received a copy of the GNU Affero General Public License
 * along with the Degage Web Application (file LICENSE.txt in the
 * distribution).  If not, see <http://www.gnu.org/licenses/>.
 */

package be.ugent.degage.db.dao;

import be.ugent.degage.db.DataAccessException;
import be.ugent.degage.db.models.Address;

/**
 * Data access object for addresses
 */
public interface AddressDAO {

    /**
     * Return the address record with the given id. Returns null when not found.
     */
    public Address getAddress(int id) throws DataAccessException;

    /**
     * Create a new address record
     */
    public Address createAddress(String country, String zip, String city, String street, String num) throws DataAccessException;

    /**
     * Update the address record with the given id.
     */
    public void updateAddress(Address address) throws DataAccessException;

    /**
     * Delete the address record with the given id.
     */
    public void deleteAddress(int addressId) throws DataAccessException;
}
